# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "contact service on clean architecture",
        "title": "contact service on clean architecture",
        "contact": {
            "name": "API Support",
            "email": "evgeniydammer@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/groups/": {
            "get": {
                "description": "Get group list method.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Get group list method.",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Item count",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Item offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "name",
                        "description": "Field sort",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/group.ListGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create group method.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Create group method.",
                "parameters": [
                    {
                        "description": "Group data",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/group.ShortGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/group.ResponseGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/groups/{id}": {
            "get": {
                "description": "Get group by ID method.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Get group by ID method.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/group.ResponseGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update group method.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Update group method.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Group data",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/group.ShortGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/group.ResponseGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete group method.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Delete group method.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "group.ListGroup": {
            "type": "object",
            "properties": {
                "limit": {
                    "description": "Limit of groups in request",
                    "type": "integer",
                    "default": 10,
                    "minimum": 0,
                    "example": 10
                },
                "list": {
                    "description": "List of groups",
                    "type": "array",
                    "minItems": 0,
                    "items": {
                        "$ref": "#/definitions/group.ResponseGroup"
                    }
                },
                "offset": {
                    "description": "Offset by groups",
                    "type": "integer",
                    "default": 0,
                    "minimum": 0,
                    "example": 20
                },
                "total": {
                    "description": "Total groups in system",
                    "type": "integer",
                    "default": 0,
                    "minimum": 0,
                    "example": 10
                }
            }
        },
        "group.ResponseGroup": {
            "type": "object",
            "required": [
                "createdAt",
                "id",
                "modifiedAt",
                "name"
            ],
            "properties": {
                "contactsAmount": {
                    "description": "Contacts count in group",
                    "type": "integer",
                    "default": 10,
                    "minimum": 0
                },
                "createdAt": {
                    "description": "Create date",
                    "type": "string"
                },
                "description": {
                    "description": "Group description",
                    "type": "string",
                    "maxLength": 1000,
                    "example": "Описание группы"
                },
                "id": {
                    "description": "Group ID",
                    "type": "string",
                    "format": "uuid",
                    "example": "00000000-0000-0000-0000-000000000000"
                },
                "modifiedAt": {
                    "description": "Update date",
                    "type": "string"
                },
                "name": {
                    "description": "Group name",
                    "type": "string",
                    "maxLength": 100,
                    "example": "Название группы"
                }
            }
        },
        "group.ShortGroup": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "description": "Group description",
                    "type": "string",
                    "maxLength": 1000,
                    "example": "Описание группы"
                },
                "name": {
                    "description": "Group name",
                    "type": "string",
                    "maxLength": 100,
                    "example": "Название группы"
                }
            }
        },
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "info": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}